# src/main/resources/contract/api.yaml
openapi: 3.0.0
info:
  title: Proyecto DAW - Empresa de Tatuajes
  description: API para manejar el back-end de una empresa de tatuajes.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
...
  /Category:
    get:
      summary: Obtener todas las categorías
      description: "Permite a cualquier usuario obtener la lista de categorías disponibles, incluyendo sus subcategorías."
      operationId: getCategories
      tags:
        - Categories
      responses:
        "200":
          description: Lista de categorías obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                title: CategoriesResponse
                properties:
                  categories:
                    type: array
                    description: Lista de categorías disponibles
                    items:
                      $ref: "#/components/schemas/Category"  # Aquí debe referenciar la clase Category
                      properties:
                        id_category:
                          type: integer
                        name:
                          type: string
                        description:
                          type: string
                        subcategories:
                          type: array
                          items:
                            $ref: "#/components/schemas/Subcategory"
        "404":
          description: No se encontraron categorías
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Crear una nueva categoría
      description: "Permite crear una nueva categoría en la base de datos."
      operationId: createCategory
      tags:
        - Categories
      requestBody:
        description: Datos de la nueva categoría
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        "201":
          description: Categoría creada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Solicitud incorrecta (Bad Request)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /Category/{idCategory}:
    get:
      summary: Obtener una categoría por ID
      description: "Permite obtener los datos de una categoría mediante su ID único."
      operationId: getCategoryById
      tags:
        - Categories
      parameters:
        - name: idCategory
          in: path
          required: true
          description: ID único de la categoría
          schema:
            type: integer
      responses:
        "200":
          description: Categoría encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Solicitud incorrecta (Bad Request)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Actualizar datos de una categoría
      description: "Permite a los usuarios autenticados con privilegios actualizar la información de una categoría. No se permite cambiar el ID de la categoría."
      operationId: updateCategory
      tags:
        - Categories
      parameters:
        - name: idCategory
          in: path
          required: true
          description: ID único de la categoría que se desea actualizar
          schema:
            type: integer
      requestBody:
        description: Nuevos datos de la categoría
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        "204":
          description: Categoría actualizada satisfactoriamente (sin contenido).
        "400":
          description: Error en la solicitud o datos incorrectos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Categoría no encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Eliminar una categoría
      description: "Permite eliminar una categoría mediante su ID."
      operationId: deleteCategory
      tags:
        - Categories
      parameters:
        - name: idCategory
          in: path
          required: true
          description: ID único de la categoría que se desea eliminar
          schema:
            type: integer
      responses:
        "200":
          description: Categoría eliminada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "400":
          description: Solicitud incorrecta (Bad Request)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id_category:
          type: integer
          description: "Identificador único de la categoría"
        name:
          type: string
          description: "Nombre de la categoría"
        description:
          type: string
          description: "Descripción de la categoría"
        subcategories:
          type: array
          description: "Lista de subcategorías asociadas a esta categoría"
          items:
            $ref: "#/components/schemas/Subcategory"
    Subcategory:
      type: object
      properties:
        id_subcategory:
          type: integer
          description: "Identificador único de la subcategoría"
        id_category:
          type: integer
          description: "Identificador de la categoría a la que pertenece esta subcategoría"
        name:
          type: string
          description: "Nombre de la subcategoría"
        description:
          type: string
          description: "Descripción de la subcategoría"




    # Productos actualizados (no tocar)
...
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensaje indicando el éxito de la operación.
      required:
        - message
...
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          description: Código del error
        errorMessage:
          type: string
          description: Descripción del error
security:
  - bearerAuth: []